+++ dart/optimizer/nlopt/NloptSolver.hpp
@@ -35,6 +35,7 @@
 
 #include <nlopt.hpp>
 
+#include "dart/common/Deprecated.hpp"
 #include "dart/optimizer/Solver.hpp"
 
 namespace dart {
@@ -42,21 +43,89 @@ namespace optimizer {
 
 class Problem;
 
-/// \brief class NloptSolver
+/// NloptSolver is a nonlinear programming solver that provides many unlerlying
+/// algorithms through nlopt (an third-party library:
+/// https://nlopt.readthedocs.io/).
+///
+/// The algorithms falls into four categories:
+///   (1) Global derivative-free
+///   (2) Global gradient-based
+///   (3) Local derivative-free
+///   (4) Local gradient-based,
+/// which can be specified by NloptSolver::Algorithm. The element of NloptSolver
+/// are mostly of the form NLOPT_{G,L}{N,D}_xxxx, where G/L denotes global/local
+/// optimization and N/D denotes derivative-free/gradient-based algorithms,
+/// respectively. For the details, please see:
+/// https://nlopt.readthedocs.io/en/latest/NLopt_Algorithms/
 class NloptSolver : public Solver
 {
 public:
+  enum Algorithm
+  {
+    GN_DIRECT = 0,
+    GN_DIRECT_L,
+    GN_DIRECT_L_RAND,
+    GN_DIRECT_NOSCAL,
+    GN_DIRECT_L_NOSCAL,
+    GN_DIRECT_L_RAND_NOSCAL,
+    GN_ORIG_DIRECT,
+    GN_ORIG_DIRECT_L,
+    GD_STOGO,
+    GD_STOGO_RAND,
+    LD_LBFGS_NOCEDAL,
+    LD_LBFGS,
+    LN_PRAXIS,
+    LD_VAR1,
+    LD_VAR2,
+    LD_TNEWTON,
+    LD_TNEWTON_RESTART,
+    LD_TNEWTON_PRECOND,
+    LD_TNEWTON_PRECOND_RESTART,
+    GN_CRS2_LM,
+    GN_MLSL,
+    GD_MLSL,
+    GN_MLSL_LDS,
+    GD_MLSL_LDS,
+    LD_MMA,
+    LN_COBYLA,
+    LN_NEWUOA,
+    LN_NEWUOA_BOUND,
+    LN_NELDERMEAD,
+    LN_SBPLX,
+    LN_AUGLAG,
+    LD_AUGLAG,
+    LN_AUGLAG_EQ,
+    LD_AUGLAG_EQ,
+    LN_BOBYQA,
+    GN_ISRES,
+    AUGLAG,
+    AUGLAG_EQ,
+    G_MLSL,
+    G_MLSL_LDS,
+    LD_SLSQP,
+    LD_CCSAQ,
+    GN_ESCH,
+    NUM_ALGORITHMS ///< Not an algorithm, just the number of them
+  };
 
   /// Default Constructor
-  NloptSolver(const Solver::Properties& _properties = Solver::Properties(),
-              nlopt::algorithm _alg = nlopt::LN_COBYLA);
+  DART_DEPRECATED(6.9)
+  NloptSolver(const Solver::Properties& properties, nlopt::algorithm alg);
+
+  /// Default Constructor
+  NloptSolver(
+      const Solver::Properties& properties = Solver::Properties(),
+      Algorithm alg = LN_COBYLA);
 
   /// Alternative Constructor
-  NloptSolver(std::shared_ptr<Problem> _problem,
-              nlopt::algorithm _alg = nlopt::LN_COBYLA);
+  DART_DEPRECATED(6.9)
+  NloptSolver(std::shared_ptr<Problem> problem, nlopt::algorithm alg);
+
+  /// Alternative Constructor
+  NloptSolver(std::shared_ptr<Problem> problem, Algorithm alg = LN_COBYLA);
 
   /// Destructor
-  virtual ~NloptSolver();
+  ~NloptSolver() override;
 
   // Documentation inherited
   bool solve() override;
@@ -71,47 +140,64 @@ public:
   std::shared_ptr<Solver> clone() const override;
 
   /// Copy the Properties of another NloptSolver
-  void copy(const NloptSolver& _other);
+  void copy(const NloptSolver& other);
 
   /// Copy the Properties of another NloptSolver
-  NloptSolver& operator=(const NloptSolver& _other);
+  NloptSolver& operator=(const NloptSolver& other);
 
   /// Set the algorithm that is to be used by the nlopt solver
-  void setAlgorithm(nlopt::algorithm _alg);
+  DART_DEPRECATED(6.9)
+  void setAlgorithm(nlopt::algorithm alg);
+
+  /// Set the algorithm that is to be used by the nlopt solver
+  void setAlgorithm(Algorithm alg);
 
   /// Get the algorithm that is to be used by the nlopt solver
+  DART_DEPRECATED(6.9)
   nlopt::algorithm getAlgorithm() const;
 
+  /// Get the algorithm that is to be used by the nlopt solver
+  Algorithm getAlgorithm2() const;
+  // TODO(JS): Rename to getAlgorithm2() once getAlgorithm() is removed in
+  // DART 7
+
 private:
-  /// \brief Wrapping function for nlopt callback function, nlopt_func
-  static double _nlopt_func(unsigned _n,
-                            const double* _x,
-                            double* _gradient,  // nullptr if not needed
-                            void* _func_data);
+  /// Converts nlopt::algorithm to NloptSolver::Algorithm
+  static nlopt::algorithm convertAlgorithm(Algorithm algorithm);
 
-  /// \brief Wrapping function for nlopt callback function, nlopt_mfunc
-  static void _nlopt_mfunc(unsigned _m,
-                           double* _result,
-                           unsigned _n,
-                           const double* _x,
-                           double* _gradient,  // nullptr if not needed
-                           void* _func_data);
+  /// Converts NloptSolver::Algorithm to nlopt::algorithm
+  static Algorithm convertAlgorithm(nlopt::algorithm algorithm);
 
-  /// \brief NLOPT data structure
+  /// Wrapping function for nlopt callback function, nlopt_func
+  static double _nlopt_func(
+      unsigned n,
+      const double* x,
+      double* gradient, // nullptr if not needed
+      void* func_data);
+
+  /// Wrapping function for nlopt callback function, nlopt_mfunc
+  static void _nlopt_mfunc(
+      unsigned m,
+      double* result,
+      unsigned n,
+      const double* x,
+      double* gradient, // nullptr if not needed
+      void* func_data);
+
+  /// NLOPT data structure
   std::unique_ptr<nlopt::opt> mOpt;
 
   /// Algorithm to be used by the nlopt::opt
   nlopt::algorithm mAlg;
 
-  /// \brief Optimization parameters
+  /// Optimization parameters
   std::vector<double> mX;
 
-  /// \brief Optimum value of the objective function
+  /// Optimum value of the objective function
   double mMinF;
 };
 
-}  // namespace optimizer
-}  // namespace dart
+} // namespace optimizer
+} // namespace dart
 
-#endif  // DART_OPTIMIZER_NLOPT_NLOPTSOLVER_HPP_
-
+#endif // DART_OPTIMIZER_NLOPT_NLOPTSOLVER_HPP_